split portable artifacts into workspace crates
