+++ Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2024"
 rust-version = "1.94.0"
 
+[workspace]
+members = [".", "crates/argent-artifact", "crates/silverscript-abi"]
+resolver = "3"
+
 [lib]
 name = "argent"
 path = "src/lib.rs"
@@ -13,13 +17,13 @@ name = "argentc"
 path = "src/main.rs"
 
 [dependencies]
+argent-artifact = { path = "crates/argent-artifact" }
 blake2b_simd = "1.0.2"
-faster-hex = "0.10"
 kaspa-consensus-core = { git = "https://github.com/kaspanet/rusty-kaspa", tag = "v2.0.1" }
 kaspa-txscript = { git = "https://github.com/kaspanet/rusty-kaspa", tag = "v2.0.1" }
 kaspa-txscript-errors = { git = "https://github.com/kaspanet/rusty-kaspa", tag = "v2.0.1" }
-serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+silverscript-abi = { path = "crates/silverscript-abi" }
 silverscript-lang = { git = "https://github.com/kaspanet/silverscript", rev = "d25bd3427a093c17327ca3d6b9e1aa5f7688c863", package = "silverscript-lang" }
 thiserror = "2"
