+++ crates/argent-artifact/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "argent-artifact"
+version = "0.1.0"
+edition = "2024"
+rust-version = "1.94.0"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+silverscript-abi = { path = "../silverscript-abi" }
+
+[dev-dependencies]
+serde_json = "1"
