+++ crates/silverscript-abi/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "silverscript-abi"
+version = "0.1.0"
+edition = "2024"
+rust-version = "1.94.0"
+
+[dependencies]
+faster-hex = "0.10"
+serde = { version = "1", features = ["derive"] }
+thiserror = "2"
+
+[dev-dependencies]
+serde_json = "1"
