+++ src/lib.rs
@@ -4,7 +4,6 @@ pub mod builder;
 pub mod codec;
 pub mod emit;
 pub mod error;
-pub(crate) mod hex;
 pub mod lexer;
 pub mod loader;
 pub mod parser;
